Started over
We determined that syslog quit recording logs from the ciscos on
12/3.  We discovered this last friday.  I git it working again
and put back the - to drop the records before they got to the
default syslog.  I need to go back and add the code to drop the
the log messages that match the ip filter so they don't get put in
the syslog that resides in the local.